MNT: explicitly check for Figure types
Sniffing for `obj.number` is actually checking for "pyplot managed figures" not
any figure (the attribute gets monkey-patched in as part of creating and
registering the FigureManger via `plt.figure`).  As matplotlib is already
unconditionally imported in this module do the isinstance check directly.

The try...except is to support matplotlib < 3.4 when we added support for
SubFigure.